% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lambda3.R
\name{lambda3.ml}
\alias{lambda3.ml}
\title{Estimate the dependence parameter of the Bivariate Poisson Distribution}
\usage{
lambda3.ml(x1, x2, mu1, mu2)
}
\arguments{
\item{x1}{non-negative integer.}

\item{x2}{non-negative integer.}

\item{mu1}{vector of non-negative parameters of the bivariate Poisson distribution.}

\item{mu2}{vector of non-negative parameters of the bivariate Poisson distribution.}
}
\description{
Given a pair of vectors of independent Poisson rate parameters mu1 and mu2,
estimate the dependece parameter lambda3 in the bivariate Poisson distribution.
}
\details{
Remember that the bivariate Poisson distribution of two random
variables x1 and x2 is defined as x1 = y1 + y3 and x2 = y2 + y3,
where y1, y2, and y3 are independent Poisson distributed random
variables with parameters lambda1, lambda2 and lambda3. Hence the bivariate
Poisson distribution would need to be parameterized as

lambda1 = mu1 - lambda3

lambda2 = mu2 - lambda3

lambda3 = lambda3
}
